[MIG] session_db to 16.0
